change total number of pagination to Int instead of Long

+++ client_end_user/shared/src/commonMain/kotlin/data/gateway/remote/BaseGateway.kt
@@ -84,7 +84,7 @@ abstract class BaseGateway(val client: HttpClient) {
     }
     fun <T> paginateData(result: List<T>, page: Int, total: Long): PaginationItems<T> {
         return PaginationItems(
-            total = total,
+            total = total.toInt(),
             page = page,
             items = result
         )
